feat(pipeline): auto-execute all commands at end of with
This is a breaking change wherein exiting the scope of the `with
pipeline()` block will automatically execute all pending commands.

For example:
```
async with await r.pipeline() as pipe:
    await pipe.delete('foo')

```

Note that `await pipe.execute()` can still be called within the pipeline
in cases where you want to retrieve the results early, run multiple sets
of commands in the same pipeline, or force exceptions to be raised
early.

For example:
```
async with await r.pipeline() as pipe:
    await pipe.delete('foo')
    res = await pipe.execute()

    if res[0] == 1:  # a key was deleted in the previous setp
        await pipe.incr('deleted-keys')
        res = await pipe.execute()
        print(f"I've deleted {res[0]} keys so far!")

```

Co-authored-by: Kevin James <kevinjames@thekev.in>