feat(pool): implement blocking connection pool
Add new BlockingConnectionPool, which can be dropped in in-place of the
default connection pool. When using the new pool, rather than raising an
exception if we attempt to make too many connections, the given command
is queued up to some timeout value instead.

Note that the BlockingConnectionPool is currently only supported for the
redis client rather than the cluster client.

Breaking change: `ConnectionPool.get_connection()` is now awaitable.

Note as well that the Pubsub client now fetches the connection's
encoding lazily prior to its first usage rather than on init.

Co-authored-by: Kevin James <kevinjames@thekev.in>