feat(compat): add support for python 3.10 (#13)
Though 3.10 is still in RC, there's a bunch of things they've already settled on which we need to tackle before being able to support it. Most noteably, a bunch of the deprecated usages of `loop` for asyncio have now been removed entirely, which means we'll need to update accordingly. Since >=3.6 supports removing this flag and 3.5 is already EOL'd, I've elected to drop support for 3.5 rather than maintain duplicate flows for all relevant code paths.

A good chunk of this PR ended up being lint fixes; the test file needed to be split since pre-commit was whinging, the linters getting updated to py3.6 where able to migrate us to f-strings, and the now-unnecessary iter* "convenience" methods could be removed.

Note that, as this removes support for 3.5, it should be considered a breaking change.